Create Content Entity
